+++ src/pages/Dashboard/Dashboard.js
@@ -24,6 +24,7 @@ import GasReceivedChart from '../../molecules/GasReceivedChart/GasReceivedChart'
 import CumulativeWaterUsageChart from '../../molecules/CumulativeWaterUsageChart/CumulativeWaterUsageChart';
 import ElectricityWeekChart from '../../molecules/ElectricityWeekChart/ElectricityWeekChart';
 import GasReceivedWeekChart from '../../molecules/GasReceivedWeekChart/GasReceivedWeekChart';
+import WaterReceivedWeekChart from '../../molecules/WaterReceivedWeekChart/WaterReceivedWeekChart';
 import ElectricityUsageMonth from '../../ElectricityUsageMonth';
 import GasUsageMonth from '../../GasUsageMonth';
 
@@ -108,6 +109,9 @@ export default function Dashboard() {
 				</Widget>
 				<Widget title="Week" name="gas-usage-current-week" icon={<FireIcon />}>
 					<GasReceivedWeekChart />
+				</Widget>
+				<Widget title="Week" name="water-usage-current-week" icon={<WaterIcon />}>
+					<WaterReceivedWeekChart />
 				</Widget>
 				<Widget title="Month" name="electrical-usage-by-month" icon={<BoltArrowDownIcon />}>
 					<ElectricityUsageMonth />